--- a/.isort.cfg
+++ b/.isort.cfg
@@ -3,3 +3,4 @@ force_alphabetical_sort_within_sections=true
 lines_after_imports=2
 profile=black
 treat_comments_as_code=true
+extra_standard_library=typing_extensions