--- a/docs/api/experimental/noinline.md
+++ b/docs/api/experimental/noinline.md
@@ -1,0 +1,6 @@
+# No-inline
+
+This offers tools to trade runtime for compile time. (Increase runtime to
+decrease compile time.)
+
+::: equinox.experimental.noinline