--- a/equinox/experimental/__init__.py
+++ b/equinox/experimental/__init__.py
@@ -1,3 +1,4 @@
 from .batch_norm import BatchNorm
+from .noinline import noinline
 from .spectral_norm import SpectralNorm
 from .stateful import get_state, set_state, StateIndex