TL;DR: XLA sub-graphs!

### Background

At present, JAX inlines the entire computation into a single XLA graph.

However, many scientific computing applications involve defining some modestly complicated function and then calling this function numerous times in different contexts. For example the vector field for an ODE must be traced 22 times when using the Kvaerno5 solver with automatic initial step size selection. (In ways that cannot easily be tidied up into a `lax.scan` or similar.)

Inlining without awareness of the repeated structure means the compiler is less efficient than it could be. I know of current examples with compile times about an hour long.

To support this use case there's been talk for quite a while about adding support to JAX or XLA for sub-programs, e.g.
https://github.com/google/jax/issues/10974
https://github.com/google/jax/issues/4572
https://github.com/google/jax/issues/3847
https://github.com/google/jax/issues/9298

### no-inline decorator

Introducing `equinox.experimental.noinline`. This decorator places the function in a separate XLA computation graph, and links it up with the main one via `jax.experimental.host_callback.call`. The decorated function is only traced once; only one copy of it exists as a jaxpr; it is only compiled once. It can still be transformed via grad, vmap etc.

Running the included benchmark `benchmarks/noinline.py` we obtain a reduction in compile time 36 seconds -> 25 seconds, at the expense of a large runtime increase, 0.002 seconds -> 0.6 seconds. In practice that's still a good net saving (36 seconds -> 25.6 seconds) in the common use-case that you're developing + debugging your program.

Going further and switching the solver in the benchmark from `dfx.Kvaerno5()` to `dfx.Dopri8()` gives even better results: a compile time reduction of 23 seconds -> 8 seconds (!), with a runtime increase of 0.002 seconds -> 0.16 seconds. (I chose not to implement this as the default benchmark, just because I have other unrelated plans for improving the compile time of `Dopri8`.)

### Limitations

1. All the awful hackery and monkey-patching of JAX internals needed to make this work.
2. This will only be efficient on the CPU. On the GPU it'll entail copies to and from the device. However I speculate that this isn't actually necessary, and may just be a limitation of our current use of `host_callback`?
3. The runtime performance has cratered. I speculate a lot of that cost is due to the back-and-forth switching via Python, again due to our use of `host_callback`. (Flame graphs TBD.) Possibly also something GIL related?

Nonetheless, our main use-case is on the CPU and the overall compile-time improvements on the benchmark represent compile speed improvements of 1.5x to 3x, which is enough to make me happy. This is something we're looking forward to relying on as those 1+ hour compile times are really biting us.

### CC

@shoyer and @YouJiacheng as I know you've both wanted this functionality in the past.
@federicov (+team) for using this.

Also speculatively tagging @gnecula (I have it in my head that you're behind `host_callback`?) @mattjj for possible interest. (Feel free to ignore this.)