@YouJiacheng 2022-07-03
I wonder why we cannot implement this functionality by register user function as primitive with help of `mlir.lower_fun` and `mlir.cache_lowering`.
Though it seems that in the MHLO=>HLO pass, `mlir.cache_lowering` wrapped function will still be expanded:
https://github.com/google/jax/blob/118db407f24f20a77ef491d504290f8c29d57d05/jax/_src/lax/lax.py#L1952-L1955