@patrick-kidger 2022-07-03
Doing so with MLIR: right! After implementing the above I started wondering something similar. As an end goal, could we arrange to generate a single `XlaComputation` that all usages point at, via XLA's `call` operation? If need be wrap it in a fake while loop (same trick `jax.checkpoint` uses) to prevent any inlining analysis that might slow things down, but I assume just having a single computation object, rather than numerous, would already be enough to improve things sufficiently already.

The implications of this would be quite dramatic: it would enable a "bounded while loop" construct -- i.e. the core backbone of diffeq solvers, nonlinear optimisers, etc. This construct requires recursively tracing nested operations of the form `lax.cond(pred, fn, jax.checkpoint(fn), *operands)`. The fact that they're nested means that at present we end up with an exponential explosion -- each use of the above involves tracing two copies of `fn`, each of which trace... etc. -- which makes this infeasible for compile-time performance. As such all current diffeq solvers / nonlinear optimisers / etc. currently hack around this and instead suffer substantial runtime performance penalties as a result. Being able to apply a `noinline` decorator to `fn` would avoid this issue, and thus drop the compile-time complexity from `O(exp n)` to just `O(n)`. This would be amazing!

That said, a `noinline` implementation like the above sounds probably simple enough (if you know what you're doing), and as per the previous paragraph this feature request is clearly important enough, that I would have thought something like the above would exist in core JAX already by now. We're really bumping up against the edge of my knowledge wrt lowerings here, and perhaps this is impossible for some reason / entirely separate XLA graphs really are necessary.