@YouJiacheng 2022-07-03
If wrap with rng fake while loop can successfully prevent inlining(really?), implement `noinline` should be easier.
But I guess this trick can only prevent computation from moving outside. XLA can still clone the code in call site.