@YouJiacheng 2022-07-03
My trial:
```python
from functools import partial, lru_cache

import jax
from jax import core
from jax.interpreters import mlir, xla, partial_eval as pe
from jax.tree_util import tree_flatten, tree_unflatten
import jax.linear_util as lu
from jax._src.api_util import flatten_fun

def foo(x):
    print('traced')
    return x + 1

@lru_cache
def abstract_eval_fun(fun, *args, **kwargs):
    args_flat, in_tree = tree_flatten((args, kwargs))
    wrapped_fun, out_tree = flatten_fun(lu.wrap_init(fun), in_tree)
    out = pe.abstract_eval_fun(wrapped_fun.call_wrapped, *args_flat)
    return tree_unflatten(out_tree(), out)

foo_p = core.Primitive('foo')
foo_p.def_impl(partial(xla.apply_primitive, foo_p))
foo_p.def_abstract_eval(lambda x: abstract_eval_fun(foo, x))
mlir.register_lowering(foo_p, mlir.cache_lowering(mlir.lower_fun(foo, multiple_results=False)))


def bar(x):
    x = foo_p.bind(x)
    x = foo_p.bind(x)
    x = foo_p.bind(x)
    x = foo_p.bind(x)
    return x

print(jax.make_jaxpr(bar)(1))
print(jax.jit(bar).lower(1).compiler_ir()) # MHLO
print(jax.jit(bar).lower(1).compile().compiler_ir()[0].to_string()) # post-compilation HLO
```
Only inlined at post-compilation HLO level
```
traced
{ lambda ; a:i32[]. let
    b:i32[] = foo a
    c:i32[] = foo b
    d:i32[] = foo c
    e:i32[] = foo d
  in (e,) }
traced
module @jit_bar.0 {
  func public @main(%arg0: tensor<i32>) -> tensor<i32> {
    %0 = call @foo(%arg0) : (tensor<i32>) -> tensor<i32>
    %1 = call @foo(%0) : (tensor<i32>) -> tensor<i32>
    %2 = call @foo(%1) : (tensor<i32>) -> tensor<i32>
    %3 = call @foo(%2) : (tensor<i32>) -> tensor<i32>
    return %3 : tensor<i32>
  }
  func private @foo(%arg0: tensor<i32>) -> tensor<i32> {
    %0 = mhlo.constant dense<1> : tensor<i32>
    %1 = mhlo.add %arg0, %0 : tensor<i32>
    return %1 : tensor<i32>
  }
}

traced
HloModule jit_bar.1

ENTRY %main.22 (Arg_0.1: s32[]) -> s32[] {
  %Arg_0.1 = s32[] parameter(0)
  %constant_7 = s32[] constant(4)
  ROOT %add.3 = s32[] add(s32[] %Arg_0.1, s32[] %constant_7), metadata={op_name="jit(bar)/jit(main)/add" source_file="*" source_line=28}
}
```